:truck: Moved openURL from layout to hooks utils
